CreatePty

+++ el_term_windows.cpp
@@ -9,8 +9,8 @@ RowCol GetTermSize() {
   GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi);
 
   return {
-      .Row = static_cast<int>(csbi.srWindow.Bottom - csbi.srWindow.Top + 1),
-      .Col = static_cast<int>(csbi.srWindow.Right - csbi.srWindow.Left + 1),
+      .Row = static_cast<short>(csbi.srWindow.Bottom - csbi.srWindow.Top + 1),
+      .Col = static_cast<short>(csbi.srWindow.Right - csbi.srWindow.Left + 1),
   };
 }
