+++ el_types.h
@@ -3,8 +3,8 @@
 namespace el {
 
 struct RowCol {
-  int Row;
-  int Col;
+  short Row;
+  short Col;
 };
 
 } // namespace el
